Only log to `STDOUT/STDERR`, not a database
Because we don't need to log to the database, we have a lot of extra code related to testing, viewing, deleting, and creating logs. Remove/alter this code so that `ActivityLog` is now a wrapper around `STDOUT/STDERR` that prints `JSON` useful messages about tasks.

These messages are handy for debugging purposes only.