Mergeable fields should be cleared on regenerate
Currently we merge fields like matching queries with whatever is there
at the moment. We want to ensure this is merged when merging successive query results,
but we need to clear it at the beginning or outdated values get preserved.

This is most visible when renaming a query and then regenerating.